requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiofiles=22.1.0=py39hca03da5_0
aiosqlite=0.18.0=py39hca03da5_0
anyio=3.5.0=py39hca03da5_0
appnope=0.1.2=py39hca03da5_1001
argon2-cffi=21.1.0=py39h5161555_0
arrow=1.2.3=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
babel=2.12.1=pyhd8ed1ab_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pyha770c72_0
bleach=6.0.0=pyhd8ed1ab_0
brotli-python=1.0.9=py39hfb83b0d_5
ca-certificates=2023.08.22=hca03da5_0
certifi=2023.7.22=py39hca03da5_0
cffi=1.15.1=py39h80987f9_3
charset-normalizer=3.2.0=pyhd8ed1ab_0
comm=0.1.4=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h458e800_1
debugpy=1.6.7=py39h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.0.4=py39hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h313beb8_0
faker=0.7.4=pypi_0
fonttools=4.43.1=pypi_0
fqdn=1.5.1=pypi_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
glib=2.69.1=h514c7bf_2
glob2=0.7=pypi_0
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
h11=0.14.0=pypi_0
icu=68.1=hc377ac9_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=py39hca03da5_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=6.1.0=pyhd8ed1ab_0
ipykernel=6.25.0=py39h33ce5c2_0
ipython=8.15.0=py39hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.1=py39hca03da5_1
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h80987f9_1
json5=0.9.14=pyhd8ed1ab_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=py39hca03da5_8
jupyter-lsp=2.2.0=pyhd8ed1ab_0
jupyter_client=7.4.9=py39hca03da5_0
jupyter_console=6.6.3=py39hca03da5_0
jupyter_core=5.3.0=py39hca03da5_0
jupyter_events=0.6.3=py39hca03da5_0
jupyter_server=1.23.4=py39hca03da5_0
jupyter_server_fileid=0.9.0=py39hca03da5_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyter_server_ydoc=0.8.0=py39hca03da5_1
jupyter_ydoc=0.2.4=py39hca03da5_0
jupyterlab=3.6.3=py39hca03da5_0
jupyterlab-server=2.25.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.5=py39hca03da5_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h8380606_1
lerc=3.0=hc377ac9_0
libclang=14.0.6=default_h1b80db6_1
libclang13=14.0.6=default_h24352ff_1
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libffi=3.4.4=hca03da5_0
libiconv=1.16=h1a28f6b_2
libllvm14=14.0.6=h7ec7a93_3
libpng=1.6.39=h80987f9_0
libpq=12.15=h449679c_1
libsodium=1.0.18=h1a28f6b_0
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h372ba2a_0
libxslt=1.1.37=habca612_0
llvm-openmp=14.0.6=hc6e5704_0
lxml=4.9.3=py39h50ffb84_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py39h1a28f6b_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=py39hca03da5_0
mistune=0.8.4=py39h1a28f6b_1000
mysql=5.7.24=he1ceea5_2
nbclassic=0.5.5=py39hca03da5_0
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=6.5.4=py39hca03da5_0
nbformat=5.9.2=pyhd8ed1ab_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=py39hca03da5_0
notebook=7.0.4=pypi_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numpy=1.25.2=pypi_0
openssl=1.1.1w=h1a28f6b_0
outcome=1.2.0=pypi_0
overrides=7.4.0=pyhd8ed1ab_0
packaging=23.1=py39hca03da5_0
pandas=2.1.0=pypi_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hc377ac9_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.1.0=pypi_0
pip=23.2.1=py39hca03da5_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=3.10.0=py39hca03da5_0
plotly=5.18.0=pypi_0
ply=3.11=py39hca03da5_0
prettytable=3.9.0=pypi_0
prometheus_client=0.17.1=pyhd8ed1ab_0
prompt-toolkit=3.0.36=py39hca03da5_0
prompt_toolkit=3.0.36=hd3eb1b0_0
psutil=5.9.0=py39h1a28f6b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py4j=0.10.9.7=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.1=py39hca03da5_1
pymupdf=1.23.8=pypi_0
pymupdfb=1.23.7=pypi_0
pyobjc-core=9.0=py39h3eb5a62_1
pyobjc-framework-cocoa=9.0=py39hb094c41_0
pyparsing=3.1.1=pypi_0
pypdf2=3.0.1=pypi_0
pyperclip=1.8.2=pypi_0
pyqt=5.15.7=py39hc377ac9_0
pyqt5-sip=12.11.0=py39hc377ac9_0
pyrsistent=0.18.0=py39h1a28f6b_0
pysocks=1.7.1=pyha2e5f31_6
pyspark=3.5.0=pypi_0
pytesseract=0.3.10=pypi_0
python=3.9.7=hc70090a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.18.0=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=5.4.1=py39h5161555_1
pyzmq=23.2.0=py39hc377ac9_0
qt-main=5.15.2=h8c8f68c_9
qt-webengine=5.15.9=h2903aaf_7
qtconsole=5.4.4=pypi_0
qtpy=2.4.0=pypi_0
qtwebkit=5.212=h19f419d_5
readline=8.2=h1a28f6b_0
referencing=0.30.2=pypi_0
requests=2.31.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.10.3=pypi_0
seaborn=0.13.0=pypi_0
selenium=4.12.0=pypi_0
send2trash=1.8.2=pyhd1c38e8_0
setuptools=68.0.0=py39hca03da5_0
sip=6.6.2=py39hc377ac9_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pyhd8ed1ab_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pyhd8ed1ab_1
sqlite=3.41.2=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
tenacity=8.2.3=pypi_0
terminado=0.17.1=pyhd1c38e8_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=hb8d0fd4_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.3.2=py39h80987f9_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py39hca03da5_0
trio=0.22.2=pypi_0
trio-websocket=0.10.4=pypi_0
typing_extensions=4.7.1=py39hca03da5_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023.3=pypi_0
undetected-chromedriver=3.5.3=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.0.4=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.6.3=pyhd8ed1ab_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py39hca03da5_0
widgetsnbextension=4.0.9=pypi_0
wsproto=1.2.0=pypi_0
xz=5.4.2=h80987f9_0
y-py=0.5.9=py39ha6e5c4f_0
yaml=0.2.5=h3422bc3_2
ypy-websocket=0.8.2=py39hca03da5_0
zeromq=4.3.4=hc377ac9_0
zipp=3.11.0=py39hca03da5_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
